Improved, encapsulated build-dir func
